requirements.txt
requests==2.31.0
google-generativeai==0.3.0
gunicorn==21.2.0
fastapi
uvicorn[standard]
python-dotenv
PyYAML
pytest
httpx
pytest-asyncio
